# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuildGraph AI Prototype API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/as-tickets/{id}:
    get:
      summary: AS ticket detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket
...
  /api/admin/agent-sessions/{id}:
    get:
      summary: Agent session detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent session
...
  /api/admin/tool-invocations/{id}:
    get:
      summary: Tool invocation detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tool invocation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolInvocation"
  /api/admin/rag-evidence/{id}:
    get:
      summary: Admin RAG evidence detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: RAG evidence detail for admin review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagEvidence"
...
  /api/admin/as-tickets/{id}:
    get:
      summary: Admin AS ticket detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Admin ticket detail with log policy and cause candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAsTicket"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ToolResult:
      type: object
      properties:
        status:
          type: string
          enum: [PASS, WARN, FAIL]
        score:
          type: number
        confidence:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        summary:
          type: string
        warnings:
          type: array
          items:
            type: string
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/RagEvidence"
...
    ToolInvocation:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        tool:
          type: string
        status:
          type: string
          enum: [PASS, WARN, FAIL]
        confidence:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        latencyMs:
          type: integer
        request:
          type: object
        result:
          $ref: "#/components/schemas/ToolResult"
    RagEvidence:
      type: object
      properties:
        id:
          type: string
        sourceId:
          type: string
        summary:
          type: string
        score:
          type: number
        usedBy:
          type: string
    AdminAsTicket:
      type: object
      properties:
        id:
          type: string
        user:
          type: string
        symptom:
          type: string
        status:
          type: string
        logRangeMinutes:
          type: integer
          example: 30
        consentRequired:
          type: boolean
        retentionDays:
          type: integer
          example: 30
        causeCandidates:
          type: array
          items:
            type: string
        upgradeCandidates:
          type: array
          items:
            type: string